Update versioning.sh
Ensure only one thing gets bumped in a commit... By precidence. 
If we choose to bump major or minor then ensure the following versioning factors are reset to 0.
If We don't have a bump set then we will assume that the build is for testing purposes... Let's keep those images tagged separately.